- CoreData를 이용해서 유저의 진척도를 가늠합니다. - 아무것도 진행하지 않았을 때(초기 상태), 닉네임까지 설정했을 때, 연동에 성공했을 때를 기점으로 앱을 시작했을 때 각자 다른 뷰를 표시합니다.
